LPD-58926 Move ManagementBar to common components
