+++ modules/apps/site/site-cms-site-initializer/src/main/resources/META-INF/resources/js/common/components/ManagementBar.tsx
@@ -10,7 +10,7 @@ import React from 'react';
 
 type Props = {
 	backURL?: string;
-	children: React.ReactNode;
+	children?: React.ReactNode;
 	title: string;
 };
